chore(DB-266): domain-file -> infra-file 모듈 위치 변경
