+++ codecov.yml
@@ -3,4 +3,4 @@ coverage:
     patch: off
     project:
       default:
-        target: 50
+        target: 10
