+++ settings.gradle
@@ -1,8 +1,9 @@
 rootProject.name = 'DongSan'
 
 include 'dongsan-external-api'
-include 'dongsan-core'
 include 'dongsan-core:domain-rds'
+include 'dongsan-core:domain-redis'
+include 'dongsan-infra:infra-file'
+
+include 'dongsan-core'
 include 'dongsan-domain:domain-rdb'
-include 'dongsan-domain:domain-redis'
-include 'dongsan-domain:domain-file'
